Splitted engine.py for issue #93
